#179 feat: 내가 만든 차량 상세 가격 추가

+++ android/MyCarchiving/app/src/main/java/com/softeer/mycarchiving/ui/myarchive/main/MyArchiveMainScreen.kt
@@ -57,7 +57,7 @@ fun MyArchiveMainScreen(
     onSelect: (Int) -> Unit,
     onMadeCarClick: () -> Unit,
     onMadeCarDetail: (MadeCarUiModel) -> Unit,
-    onMadeCarDelete: (Int) -> Unit,
+    onMadeCarDelete: (Long) -> Unit,
     onSavedCarClick: () -> Unit,
     onSavedCarDelete: (Int) -> Unit,
 ) {
